New command for importing a single domain from a file
The command is used as follows:

    importDomainWithSettingsXML <file path> [overwrite]

It reads the file given as first argument (path must be absolute) and imports
the domain found in it.

It does not check whether it overlaps with another domain (same elements in two
domains). It does, however, check if there is an existing domain with the same
name as the one being imported. If so and if the (optional) second argument
is "overwrite", it will first delete the existing domain; if not it will refuse
to import the new domain.

Change-Id: I7aaa225f2f1a296f52dc8a97f55e1ff70c06d567
Signed-off-by: David Wagner <david.wagner@intel.com>