xmlGenerator: new generator using the python bindings
This new script (domainGenerator.py) uses both the newly-refactored
EddParser.py module as well as the new python bindings to generate XML Settings
through the parameter-framework interface.

This script can take the following data as input:

 - the toplevel config file (mandatory)
 - the 'criteria' file (mandatory)
 - an initial settings file to be amended (optional)
 - zero, one, or several 'domain' files (each containing a single domain)
 - zero, one, or several files in the EDD syntax (aka ".pfw" files)

Call the script with the "--help" argument for detailed help.

It implements its own derived class of PfwBaseTranslator, calling the
parameter-framework python bindings. Any error happening within the binding or
below (e.g. the pfw core) is collected and forwarded to the user.

Change-Id: Id11123e083ed2a487800e737007a99541949a4bb
Signed-off-by: David Wagner <david.wagner@intel.com>