xmlGenerator: Refactor PFWScriptGenerator
The existing code is split in two parts:

- "EDD"-files parsing;
- parameter-framework remote commands generation

The first part is mostly kept as-is but the file is renamed to "EddParser.py";
the second part is refactored and an object presenting a "Translator" intreface
is passed to the parsed objects in order to generate the remote
parameter-framework commands. Later, another class implementing the Translator
interface will be implemented using the python bindings.

Some dead code is removed in the process.

Change-Id: I9725600ce34f36742c7e27ea7aee53892dd799b0
Signed-off-by: David Wagner <david.wagner@intel.com>