Create a ASSERT_ALWAYS macro as utility macro.
When using the assert() function, the condition is not evaluated in
release mode, only in debug mode.

This new macro implements a condition verification that is build
type independent (active both in debug and release modes).
Add the file,function and line in error output. Using terminate()
instead of abort() allow to handle the assert. Using __FUNCTION__
instead of __func__ to be compliant with all compilers (visual
studio 2013 and upper and linux), also make the C++ file name.

Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>
Signed-off-by: Louis Le Gall <louis.le.gall@intel.com>