ParameterMgr: add a method to forcefully disable the remote interface
In some setup, users might not want to start the remote interface even if the
toplevel configuration file allows it.

The parameter-framework client can now override the remote interface starting
policy. This was, until now, only dictated by the 'TuningAllowed' attribute in
the toplevel configuration file and the presence of the libremote-processor
library.

This method is forwarded to the connectors and the bindings.

Change-Id: Ib6dc272dfc7114125fdafd1a58642cde88847752
Signed-off-by: David Wagner <david.wagner@intel.com>