Allows exportDomainsWithSettingsXML command with no file path
Currently, the exportDomainsWIthSettingsXML command expects the file path
to be passed as parameter, as an absolute path.

The default file path mentioned in ParameterFrameworkConfiguration.xml file
is used if there is no command argument. If this path is a relative one,
it is converted into absolute path before being provided to
ParameterManager.

Signed-off-by: Thierry Quervel <thierry.quervel@intel.com>