Parameter manager multithreading usage
The libxml2 library, which is a dependency of
the parameter framework, requires a specific usage
in case of multithreading.

Indeed, if the libxml2 library is used by several
threads, the function xmlInitParser() has to be called
at program start in the main thread (i.e. a thread that
lives until process exit).

This patch introduces a static object that call
xmlInitParser() before main() start.

Signed-off-by: Thomas Cahuzac <thomasx.cahuzac@intel.com>