Let a domain decide when he should be synchronized.
Domains are browsed twice when applying configurations.
The first one gather syncer of domains which are not
"sequence aware" and launch a sync on them.
The second one launch an apply without providing a syncer set.
It lets sequence aware domains to synchronize if needed.

This patch allows to browse domains only once. If a domain
is sequence aware, it will be synchronized during configurations
restauration. If not, the synchronisation will be delayed at
the end of all applies.

Sequence aware domains are now synchronized first.

Signed-off-by: Jules Clero <julesx.clero@intel.com>