Install a recent gcc, gcov, cmake and ctest version
The c pfw interface (following patch) needs a recent gcc to compile.
Gcc and gcov version need to be align, thus gcov is updated too.
Ctest analyse the gcov output, as a result it must be align with
it (gcov 2.7 changed it's report format).
Ctest, and cmake are packaged together.

Conclusion: update cmake and gcc packages.
Unfortunatly ubuntu backport does not provide a recent enough
version, that is why ppa were used instead.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>