Add functionnal tests to the pfw.
This patch adds functionnal test to the PFW.
The execution of these tests needs the test-subsystem to be compiled.

The tests can be launch with 'make test'.
We set the environment vars in the cmake.
Then we launch ACTCampaignEngine.py.

Temporary files like build time conf files are placed on the build directory.

Next steps :
- Test-subsystem needs to be reworked (with binding python).
- All tests seems to pass. Previously there were expected failures on
  some tests. For those who know these failures, please investigate on
  this "problem".

Signed-off-by: Adrien M <adrienx.martin@intel.com>