travis: run 'ldconfig' after 'sudo make install'
In 327b1a48e2fe539c0ca5414ce9c98b2405536a09 we changed the installation prefix
to /usr because we witnessed errors during unit tests using the python
bindings.

We initially thought that /usr/local/lib wasn't in the loader's default paths
but it turns out it is. In fact, it only is the loader's cache that was
outdated and needed to be regenerated. That's what ldconfig does.

Signed-off-by: David Wagner <david.wagner@intel.com>