Remove Strategy Design pattern use for Criterion type
Criterion type (Inclusive or Exclusive) was determined
thanks to Strategy pattern through SelectionCriterionType
class.
It leads to an overhead for the client who need to create
the criterion type and then the criterion itself. The client
can also delete the criterion type which can lead to hard to
debug bugs.
This patch replaces this pattern by inheritance. We now have
two criterion types, SelectionCriterion which is the basic
Exclusive criterion and Inclusive whose specialize it..

This patch has some client API impact:
    - ISelectionCriterionTypeInterface has been merged with
      ISelectionCriterionInterface
    - isTypeInclusive has been renamed isInclusive
    - getFormattedState does not take any argument any more

Signed-off-by: Jules Clero <julesx.clero@intel.com>