Fix hidden overloaded virtual function in BitParameter.h
accessAsInteger function prototype is defined in BaseParameter
abstract class with an uint32_t type. Nevertheless in BitParameter
the overloaded function take an uin64_t. It seems that GCC accepts
wrongly this mistake but clang don't.
This patch corrects the function prototype accordingly to the one
defined in BaseParameter. It allows the project to be compiled
with clang.

Change-Id: I36e9269313f34af870808b7140f0eab3cdfd9820
Signed-off-by: Jules Clero <julesx.clero@intel.com>