Add clang compilation to travis configuration
Swig has been updated in order to avoid the use of
Wno-unused-but-set-variable flag which is not present in Clang.
The coverage is not supported for now due to gcov compatibility problems
with clang.
The debug build his also built in a different directory to avoid an
empty error of python bindings with clang.

Signed-off-by: Jules Clero <julesx.clero@intel.com>