Change blackboard write/readString prototype
Those functions are used in one place and let the caller transform a
std::string to a char * which makes the use of alloca necessary.
Moreover, the readString cannot know if the pointer given in input
points to a buffer with enough space, making it dangerous.

This patch reworks the API so that write/readString() get std::string
and lets the implementation play with char *. This also allows to get
rid of alloca() use which is quite dangerous.

Change-Id: I973e4987278289bb914c4e194b86e924a1c88a6d
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>