[Travis] No longer install debug build
Debug build used to be installed as it was required
for make test.

As make test is now stand alone, do not install debug builds.
This has several advantages:
 - slightly faster builds.
 - make test without install is actually tested by travis.

Debug binaries used to overwrite release ones that were installed
presciently. As it is no longer the case, and to avoid tests to use
installed binaries instead of the local ones, test the debug build
*before* the release one.
As the debug build is now build (first) in it's own folder,
debug and release build can not interfere.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>