travis: Use the docker infrastructure
As advertise by travis news:

    http://blog.travis-ci.com/2014-12-17-faster-builds-with-container-based-infrastructure/

using the new docker infra is faster and more reliable. This comes with the
cost of not being able to use "sudo", though. Installing packages is still
possible by specifying them in an "addons.apt" section in .travis.yml. We have
to remove "sudo make install" but we aren't using the installed libs and
binaries anyway.

I had some syntax issues with the "addons" and "env" sections and had to remove
the coverity subsection which wasn't configure for the 01org repository anyway.

Signed-off-by: David Wagner <david.wagner@intel.com>