libcriterion: Rework Inclusive Criterion numerical value definition
Previously, numerical values of inclusive and standard (exclusive)
criterion was not handled in the same way.
Exclusive criterion can have any integer as numerical value. On
the contrary, inclusive criterion can only have a numerical state
which is a power of 2. It comes from the fact that a bit mask is
internally used to handle inclusive criterion state.

This implementation forces the client to register criteria with
a mask where only one bit is set.
For example, to register a literal 'A' in the third position of
a bit mask we have to do criterion.addValuePair('A', 1 << 3).
This is an error prone interface because it is not consistent
with the exclusive criterion interface.

This patch rework this interface to register the literal value
with the desired bit mask index (e.g: criterion.addValuePair('A', 3)).

Signed-off-by: Jules Clero <julesx.clero@intel.com>