xmlserializer: unify string and file sink/sources
Instead of having one class for import from a file and one class for import
from a string, only have one class for Xml Sources and wrap its creation from
either a string or a file in a helper static function.

As for export, unify export to string and to file in a single class that
exports to a stream.

Since the public APIs take:

- a string as argument that can either contain an actual XML document or a
  filename;
- a boolean stating which of the above cases it is,

we need wrappers to convert that semantic into the new "everything is a stream"
semantic. A lot of code can then be factorized.

Change-Id: I357d49fc35f98776df9518e4ea76922891a9ad9a
Signed-off-by: David Wagner <david.wagner@intel.com>