Drop the absolute path requirement on xml import/export
It made little sense to restrict importing/exporting settings to absolute file
paths.

However, it avoided mistakes of users of the remote interface who may believe
that file paths are relative to their own location whereas they are relative to
the Parameter Framework's client's working directory.

For this reason, make it clear in the remote commands's help what the
requirements on the file paths are.

Signed-off-by: David Wagner <david.wagner@intel.com>