libcriterion: Rework Value handling
In previous API, criterion values were added through addValuePair
interface. This API was error prone because we were able to create
bad criterion (with no value for example).

Moreover, the default criterion state was '0' which is not always
a valid state for an exclusive criterion which could leads to some
problems at startup.

The new criterion creation API requires that the client provide a
Values object which contains pairs the criterion will deal with.
Some check are done on this object to insure the criterion will
be valid. The default state is the first one added to the Values
object.

Signed-off-by: Jules Clero <julesx.clero@intel.com>