android/python bindings: force gcc and 32bits compilation
Since Android only provides a 32bits version of the python interpreter and
libs, we need to compile for 32bits even on 64bits hosts.

Also, since it seems that clang is the default compiler for host, we need to
explicitely forbid it as long as the parameter-framework is compiled with gcc.

Change-Id: I5d484c0e58c304680192f443acd81258de179e52
Signed-off-by: David Wagner <david.wagner@intel.com>