libcriterion: Remove getNumericalValue specialization for inclusive criterion
Inclusive criterion has their getNumericalValue specialized in order
to format a bit mask corresponding to the literal value given in
argument. This value can contains several '|' to combine values
which will composed the returned mask.
This is not consistent with its twin method getLiteralValue which
can only take an 'atomic' numerical value in parameter. Moreover,
getNumericalValue is not returning a 'numerical value' but rather a
'criterion state'.

This patch removes getNumericalValue specialization for inclusive
criterion. The method called is now the one implemented in Criterion
class and has the same behaviour.

Signed-off-by: Jules Clero <julesx.clero@intel.com>