xmlserializer: Introduce Xml binding powered serializer
Current serialization implementation require to map the design to
the desired xml files. Thus, xml can't be easily removed and the code is
hard to test.
This patch introduces a serializer, based on the current xmlSerializer
library. This serializer uses an Xml bindings structure which describe
the mapping of the data holder to the desired xml files.

Signed-off-by: Jules Clero <julesx.clero@intel.com>