Android: let the build system decide how to build C++
Let it decide which STL is to be used for both host and target binaries; at the
time of writing, it defaults to clang's stl.

Let it decide which compiler is to be used on host; at the time of writing, it
defaults to clang. Force usage of clang on target because it defaults to gcc.

Change-Id: Ib4fb8e58e9db48e27d479d47bbdb802b6e52e0d4
Signed-off-by: David Wagner <david.wagner@intel.com>

squashme: opt in for clang on target

Change-Id: I4d92c94dbdcd5142509fb7a05a6446bf3ac55ce1
Signed-off-by: David Wagner <david.wagner@intel.com>