xmlserializer: remove unused or redundant ctors and unify their API
Some constructors weren't used so they can be removed altogether. Some others
were reduced versions (fewer arguments) of other ones and can be removed and
instead have a single "full blown" constructor with default values.

This required a slight change in the order of the arguments and was the
occasion to make them all have the same look'n'feel.

Change-Id: Ifc7c1e9def3880bbdeccac9f8735c47947807069
Signed-off-by: David Wagner <david.wagner@intel.com>