libcriterion: Remove litteral/numerical value handling
Current criterion interface allows user to register a pair
of numerical/literal value. Then the user must use numerical
values to set the desired state.

This patch removes this possibility. The user is now registering
only literal values which are used to set the desired state.
The mapping between numerical and literal value, if needed, should
now be handled by the client.

Moreover, Values are now given by the client through a list not a set.
It allows the client to have the control of the default state which is
the first value of the list. Internally, we still use a set to avoid
value duplication.

Signed-off-by: Jules Clero <julesx.clero@intel.com>