doxygen: don't add timestamps to html pages
The version number and the git history of the 'gh-pages' branche are
sufficient; moreover, this unnecessarily create new objects in the git database
each time the documentation is updated. Since 'git clone' will fetch the
'gh-pages' branch along with the master branch and since this branch is quite
heavy already, it increases the time to clone the parameter-framework for
little reason.

Signed-off-by: David Wagner <david.wagner@intel.com>