Client Simulator: Allow to launch several scenarios by command line
Client Simulator only supports one scenario to be launched by
command line option.

For automation purpose, this patch adds an option allowing the user
to specify several scenarios on command line.

Also, the default behaviour used to be: "after a scenario has been executed,
present the user with a menu in order to run more of them". The new behaviour
is to exit the Client Simulator after the scenarios have been run; the old
behaviour can be reproduced with the new "--no-exit" option.

If neither --scenario nor --interactive nor --no-exit are passed, then the same
behaviour as with --no-exit is applied.

Change-Id: I0ea64a2576c390d0b80b825b9f3f320ce9e8da43
Signed-off-by: David Pertuze <davidx.pertuze@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>