Android: Make libxmserializer dynamic
The PFw as recently removed (this trick was not working on osx)
it's *_include libraries that were used to include headers
without linking against the libraries in the android build system.

There is to way to include the headers:
 - copy the headers (but that would be in contradiction with how the main pfw headers are included)
 - link against the library (chosen solution)

In order for the plugins to avoid linking against a static
libxmlserializer - ie add it and it's dependency to all plugins -
the xmlserializer library is now dynamic.

Change-Id: I79feb434c7b44ad4473d62276efc3c4b77928894