Android: Make libxmserializer dynamic
The Paramter Framework has recently removed its *_include libraries
(this trick was not working on osx). They were used to include headers
without linking against the libraries in the android build system.

There are two ways to include library headers:
 - copy the headers (but that would be in contradiction with how the main pfw headers are included)
 - link against the library (chosen solution)

In order for the plugins to avoid linking against a static
libxmlserializer - ie add it and it's dependency to all plugins -
the xmlserializer library is now dynamic.

Change-Id: I79feb434c7b44ad4473d62276efc3c4b77928894