Port C++14's make_unique
std::make_unique is a function that alocates an object and
wraps it in a std::unique_ptr. This helper function
is very useful for unique_ptr manibulation.
Unfortunately the proposal missed the C++11 time frame and
was adopted in C++14.

As is implementation is very straightforward, port it in the
PF.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>