Client Simulator: remove the tools' paths from the config items
remote-process and test-platform should simply be in the PATH and so, there is
no need for their paths to be put in conf.json. This change is somewhat
backward-compatible because these configuration items will simply be ignored.

Another reason why it wasn't very useful for these paths to be in conf.json is
that, even if their path is know, one still has to add their library's path to
LD_LIBRARY_PATH. It is easier to install the Parameter Framework project, in
order to have everything set.

Signed-off-by: David Wagner <david.wagner@intel.com>