Refined rogue element detection
CConfigurableElement::isRogue() method was only suitable for parameters,
not for elements. If an element did not belong to any domain, but one of
its descendent did, this method would have mistakenly returned true.

This error was not detected up to now for this method was only used with
parameters (leaf elements in parameter tree).

Signed-off-by: Patrick Benavoli <patrick.benavoli@intel.com>
Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>