Define special case for convertTo<uint8_t>() and convertTo<int8_t>()
The generic version was converting uint8 as it was a character
(uint8_t is an alias to unsigned char on most compiler).
Thus converting "1" would return 49 ie '1'.

As convertTo is thought as an _numerical_ conversion tool
(contrary to boost::lexical_cast for example),
forbid considering the input as a character and consider uint8_t
(aka unsigned char) as a number exclusively

Same for int8_t.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>
Signed-off-by: David Wagner <david.wagner@intel.com>