RFC: modify the criterion APIs to take uints instead of ints
Using signed integers in the criterion API made it necessary to add guards for
undefined behaviour occuring when bit-shifting a signed integer up to its sign
bit. It also artifically reduced by 1 the number of possible inclusive values
in an inclusive criterion.

All APIs are changed in a retrocompatible fashion to take unsigned ints as
input parameters. However, getNumericalValue takes an output parameter. This
one must be kept as it is. We add a overloaded version of getNumericalValue
taking an unsigned int and depracate the "signed int" version.

Users that do not update their code to pass unsigned integers will still work
fine but will have the "31 inclusive criterion values" limitation.

Signed-off-by: David Wagner <david.wagner@intel.com>