functional tests: allow TuningMode to be set
The setTuningMode() method was correctly wrapped but the template for the
toplevel configuration file was forbidding the tuning mode to be allowed. This
was done in order to avoid the remote interface to be started.

The template now set TuningAllowed to true but the ParameterFramework wrapper
class also calls setForceNoRemoteInterface(true) in order not to start the
remote interface.

Signed-off-by: David Wagner <david.wagner@intel.com>