Added a getXmlElementName() method to CElement
Up to now, Xml exporting feature in CElement relied on the
CElement::getKind() method to get the XML element tag.
However, some CElement derived classes might need to reflect in XML
as a type which differs from their kind (which is class name related).
This patch uncouples the notion of element kind from the tag used to
import/export an element in XML. Then it makes use of this functionality
each time an XML element is created based on the element XML type.

Signed-off-by: Patrick Benavoli <patrick.benavoli@intel.com>