Wrong XML "Name" attribute in enum value pairs
When exporting an element's structure (via the C++ API or via the tuning
interface), a spurious "Name" attribute appears for EnumPair nodes, while
only "Literal" and "Numerical" attributes should be present. This is because
the XML import / export feature assummes the existence of a Name attribute
for all XML nodes. The Name default processing can however be skipped from the
toXml() methods by calling the base class' childrenToXml() method instead
of the toXml() one.

Signed-off-by: Patrick Benavoli <patrick.benavoli@intel.com>