Replace 'and' by && as windows complains
For some reason, the MS compiler deactivated the possibility to use 'and'
instead of &&.  This makes the code fail to compile on windows.

This patch replaces the 'and's by && (like in legacy C).  It should be
considered as a bad practice to use 'and' in the code as it breaks
multi-plateform support.

Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>