Fixed element XML access feature
Calling base implementation of serializeXmlSettings() method from
ArrayParameter was causing the update of the XML node by Base parameter,
that is, the scalar implementation.
This patch ensures only the array implementation of the
serializeXmlSettings() is solicited.
This is achieved by a proper overloading of the doGetValue() and
doSetValue() virtual methods called from the base implementation in
BaseParameter.
Note: this issue was first introduced by the attempt of reusing base code
during XML export to take advantage of Name attribute handling. See patch
"Make XML export settings reusable for any configurable element".

Change-Id: If535ca4c4b608cf7a85ab2a90121e2ea64d00820
Signed-off-by: Patrick Benavoli <patrick.benavoli@intel.com>
Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>