Normalize unused parameter declaration
When a function parameter is unused, the compilator emits a warning.
There are several wais to avoid that,
and thus express that this parameter is rightfully unused.

 1) remove the unused parameter warning
    => but their are errors when that would be hided
 2) declare de parameter as __attribute__((unused))
    => but it is gcc/clang specific
 3) use the parameter in a no-op operation
    (usually by casting it to void)
    => but using the parameter declared as unused by this mean
       will not trigger an error
 4) remove the parameter identifier (leaving the type) from the function
    => but when the name is non redundant with the type,
       this looses information and thus make review harder.
 5) comment the parameter identifier
    => no drawback

As a result it has been chosen to normalize all unused parameters
with the 5th solution.
Some unused parameter have been renamed if they were too long.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>