Wrong type comparaison of file types
The code was converting a `pos_type` to a `size_t` to compare
a file size and a structure size. This leaded to precision loss on
windows.

`istream::pos_type` represent the position in a file stream.
Thus when at the end of the file, it represent it's size.
As a result `pos_type` must be able to represent any file size.

`size_t` on the other hand must be able to represent any contiguous
memory array size.

Those two type are not necessary the same but it is safe to assume
that `pos_type` is equal or bigger than `size_t`, as supported
system can all have files a lot bigger than the usual memory size.

As a result when comparing a file size and a structure size, use
`pos_type`.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>