C bindings: NONNULL & USERETURN windows support
The NONNULL and USERETURN macros are used in the c bindings to
try to detect api misused and raise a warning.

As this is not possible using standard C++, the code was
using g++/clang++ specific extension. As a result it
did not compile on windows.

Use the equivalent flag for USERETURN under visual studio cl
(_CheckReturn).
NONNULL does not have an equivalent on VS thus disable the macro
on this platform.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>