Use std algorithms for blackboard access
Blackboard access was implemented using `memcpy` with issues
a warning on VS13 cl.

Use standard C++ algorithm instead. This does not solve completely
the problem as their is still some warning raised because
the iterator are unsafe.

It might be wise in a future rework to use safe containers for the
blackboard in order to get rid of this warning definitively.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>