remote-processor: limit the scope of command handler object to process function
The command handler was passed at build of the object when actually it is needed
only on the processing.

This patch changes the command handler to be passed on process. This allows caller
to separate seting up a server and handling commands: the handler is created when
the server is ready.

Tracked-On:
Change-Id: I29d45302d74652c3585ec8a28ba5696c10747b13
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>