rework remoteProcessor
The command handler was passed at build of the object when actually it is needed
only on the processing.

This patch changes the command handler to be passed on process. This allows caller
to separate seting up a server and handling commands: the handler is created when
the server is ready.

Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>