rework remoteProcessor
The command handler was passed at build of the object when actually it is needed
only on the processing.

This patch changes the command handler to be passed on process. This allows caller
to separate seting up a server and handling commands: the handler is created when
the server is ready.

Tracked-On:
Signed-off-by: Sebastien Gonzalve <sebastien.gonzalve@intel.com>
## Don't forget to add a JIRA URL to the Tracked-On metadata!
## https://wiki.ith.intel.com/display/CACTUS/Patch+requirements+check
## url should be in the form:
## Tracked-On: https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
## Several Tracked-On lines are allowed
## you can fully prefill your Tracked-On commits using environment variable:
## export CURRENT_TRACKED_ON=https://jira01.devtools.intel.com/browse/<PROJ>-<NUMB>
##