Use checked iterator on windows
Visual studio raises a warning if the check iterator feature is activated
but a raw pointer is used as iterator (as it can not check it's bounds).

As it is a safety feature, do not silent the warning, but use the
microsoft specific `make_check_array_iterator` that take a pointer
and the size of the underline buffer.
For other compiler, use the raw pointer.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>