Use size_t for size containing variables
Sizes should not be represented as uint32_t as
this is non portable on 64 bit architectures.
It also raises warnings when narrowing size_t to uint32_t
on 64 bit architectures.

Change all size variables to be size_t.
The only exception is nework code as the protocol
should not be architecture dependent.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>