CMake: rely on CMAKE_PREFIX_PATH for finding user-provided dependencies
I used to think that CMAKE_PREFIX_PATH should not be set by the user (the one
running the "cmake" commands) but it turns out, it is the correct way to tell
CMake where to find user-provided packages, includes, libs, etc.

In order to set CMAKE_PREFIX_PATH to multiple paths, it can be set to a
semicolon-separated list of values (this is the case for a lot of other
variables in CMake).

Signed-off-by: David Wagner <david.wagner@intel.com>