Disambiguate make_unique in case it is provided by the STL
std::make_unique is a C++14 addition but we are, for now, sticking with the
C++11 standard. Since we have defined our own make_unique function, we need to
disambiguate it for MSVC, which defines std::make_unique.

This is done by putting our make_unique in the 'utility' namespace.

Signed-off-by: David Wagner <david.wagner@intel.com>