Rework parameter handle access check
Set parameter needs to check that the array size requested
is the same as the parameter size.

But get parameter returns all the array, as a result
there is not array size to check.

The problem was that a single function was used for
get and set access check.

Thus when wanting to check get access on an array,
no size could be given and -1 was used instead.

As a result -1 was implicitly cast to uint32_t.
Which raised a warning on windows.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>