Revert "Merge pull request #225 from dawagner/travis-no-container-based"
Travis was reverted to it's legacy mode, building with a vm instead of
a docker based container.

This change was made to be able to use sudo to install a recent cmake
package.

Nevertheless it had several disadvantages:
 - Travis.yml is less readable
 - build is slower (not that much though)
 - recipe require sudo

After double checking, a few ppa white-listed for the container
contained a cmake recent backport. So using docker Travis mode
is possible.

Revert to container build as this is the preferred and recommended
Travis build mode.

This reverts commit c094250e7d340b24f023844d516b55cbd96913ad, reversing
changes made to 7a7987c57dc674b39b6a2f3e7e37cacfaf238270.