Functional test: Use vector instead of C array
`initializer_list` does not support the move semantic.
Thus the test vectors can not be stored in a stl container
if they are non copyable.

This is why the test vector implementation was a C array.

Unfortunately VS 2013 does not support deducing the array
size from the initializer if the type is declared through an alias.

As all test vector are currently copyable, use a stl container instead.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>