Suppress warning of exported variable without a dll interface
Suppress warning 4251 (related to exported symbol without dll interface)
as they referee to private member (coding style forbids exposing attribute)
and thus are not to be used by the client.

This warning was spamming the build log and hiding potentially more
useful warnings.

A better fix would be to export only public methods
instead of the whole class, but they are too many to do that.
The canonical way of fixing such problem
is to have a separated plugin interface.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>