Value initialization is broken in VS 2013
`new T();` is supposed to zero-initialize then default initialize T.
See http://en.cppreference.com/w/cpp/language/value_initialization.

This was not the case on VS 2013 in debug thus a pointer was set to an
invalid value instead of nullptr.

Fix by making explicit that the default value of the pointer should be
nullptr.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>