WIP: only export symbols that are part of the API
"private virtual" methods need to be exported because the plugins will need to
access them at runtime.

We can't do this selective export of classes that are to be derived by plugins
(i.e. Subsystem, SubsystemObject and FormattedSubsystemObject):
For some reason I don't understand, the plugins require the "typeinfo" symbol
of these classes at runtime. Disassembling the code show that this symbol isn't
even used in the generated code but merely required by the dynamic linker...

Strangely, this increases the size of the ".text" section of the generated
library (not much but this is not what I expected nevertheless).....

Signed-off-by: David Wagner <david.wagner@intel.com>