Fix Klocwork false positive errors
KW tool has a rigid behavior regarding output parameters
if they are given unitialized. Even if it is up to the caller to check
the returned error of functions taking output parameters to ensure these
parameters are valid or not, KW still complains.

This patch workarounds these false positive errors.

Change-Id: I03e7f2b72094ce666d1b73dd6d3ee1b2a05c4bf5
Signed-off-by: François Gaffie <francois.gaffie@intel.com>