SubsystemObject: properly wrap the logger object with log methods
Plugins had access to a protected member that allowed them to log using:

    _logger.info() << "some log";

This is a design flaw because members shouldn't be exposed (either public or
protected). Instead, SubsystemObject now has two protected methods: info() and
warning() that wrap this _logger member, which is now private. Also, the code
required to log is nicer:

    info() << "some log";

Signed-off-by: David Wagner <david.wagner@intel.com>