Legacy tests: Do not use files to split a string in lines
Lots of test need to compare two list, one list being retriving as
a new line separated string.
To split this string in a list, the test code was:
 - opening a temporary file in write mode
 - writing the string in it
 - closing it
 - reopening in read mode
 - using the `readline` function of the `File` class to read each line
 - close the file
 - remove it

All this can be simplify by calling splitlines on the original string.

There arelots of other places were this refactor could be usefull,
but those test are legacy so this patch only refactor code
that failed on windows to improve the error message.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>