Legacy Tests: listCriteria takes a format parameter
The listCriteria function can either take:
 - no argument => default format
 - "XML" or "CSV" => xml or csv format

The tests were calling the command with an empty argument
wich was removed in shell mode.
In order to support non shell popen mode, correct the command.

XML (instead of default or csv format) was chosen to be tested as
this is the only used.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>