Legacy tests: Detect and log errors in commands
The code was testing that span processes succeeded by testing that err
was None.
Nevertheless stderr was not piped thus *always* None.
Thus all tests that assert that the command succeeded are dead.

Correctly captured stderr and test returncode to check that the command
succeeded.

Unfortunately lots are test are bugged and will fail if errors are not
hidden. Ie: some tests are failing but nobody was aware of it.

For now only log the error. Proper test fix will need to be done.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>