Legacy tests: Explict command that should fail
Remote command results were ignored. This leaded to hard to
debug errors.

For each command, require a success (implicit as it's the default),
or a failure (explicit `requireSuccess=False`).

A few command result were not determinable from the call frame.
As a workaround, set requireSuccess to None to delegate the
check to the caller.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>