Legacy tests: Try to avoid port conflict
Test-platform was using the 5001 port,
nevertheless this port was used on windows by mcafee.

Change to a port that is not registered by IANA and
not in the dynamic range.

Note: choosing a port in the dynamic range (49152–65535)
creates a race were the test-platform can not start and
$ ss -a |grep 50000
Netid  State      Recv-Q Send-Q   Local Address:Port       Peer Address:Port
tcp    TIME-WAIT  0      0            127.0.0.1:50000 127.0.0.1:50000

The root cause of this could not be found.
Strong correlation with the time test-platform starts has been observed.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>