Warn when loading a subsystem in virtual mode
Starting a subsystem in virtual mode is often a temporary
hack in order to start the parameter framework with a buggy
subsystem.

This modification although useful in some case,
may lead to hard to debug bug if the user is not
aware of the virtual state.

Notify the user by warning when a subsystem in loaded
in virtual.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>