appveyor: do not pollute PATH with test-only paths
We used to add the libxml2 directory to the PATH when building although we only
intended to do so for the purpose of the tests. It led CMake to find libxml2
for the wrong reasons. It actually shouldn't have been able to find it since
libxml2's dir wasn't added to the CMAKE_PREFIX_PATH.

The directory where libxml2 has been extracted is now added to the prefix path
and its 'bin' subdirectory also added to a variable only used during the tests.

Signed-off-by: David Wagner <david.wagner@intel.com>