Print the file line and column number context in libxml2 error messages
This used to be done by libxml2 itself when using the "generic error handler"
but was lost when switching to the "structured error handler".

Unfortunately, we do not have as much context as with the generic handler,
which used to also print the incriminated content and point to the exact
location of the error. I didn't find any easy way to get it.

Signed-off-by: David Wagner <david.wagner@intel.com>