ParameterHandle should handle non const elements
Handles have non const method, thus their reference
on PF elements should not be const.

It used to work because the only method called `access`
was const even when setting a value.

Give a non const value to ParameterHandle constructor.
It will be used in a future Handle that will call non
const methods.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>