Create a test toolbox with catch and tmpfile
Also, this removes TmpFile from pfw_utility which in turn was part of
libparameter. TmpFile uses unsecure functions for creating temporary files.
While this is not an issue for test purposes, it can be a security issue
elsewhere.

Even though TmpFile was only used for test purposes, having it part of
libparameter raises security warnings in static code analysers; removing it
from libparameter removes any chance of using it.

Signed-off-by: David Wagner <david.wagner@intel.com>