Serialization context is element dependant
The serialization context in exportElementToXMLString was always
a CXmlSerializingContext. Nevertheless in a lot of called method
it was wrongfully down casted to other subclass.

Until recently it did not trigger UB as exportElementToXMLString
was luckily not called on element that did the downcast.

Recently, ConfigurableElement was change to also do this down cast,
triggering a bug on export and import structure.

Change exportElementToXMLString so that it is the caller that creates
the context based on the element type.

Signed-off-by: Kevin Rocard <kevin.rocard@intel.com>