Mapping: replace an assert with an error message
The assert was checking that the mapping key being added to an element's
mapping data was not already present. It seemed redundant with the subsequent
check that was returning an error if the name of the mapping key being added
was already present.

Returning an error to the user leads avoid crashing and produces clearer error
messages.

Signed-off-by: David Wagner <david.wagner@intel.com>